fix(report-export): address PR #139 review findings
CRITICAL fixes:
- Remove shell=True from subprocess.Popen (security)
- Add MAX_FLOORS=500 ceiling to unbounded while loops
- Narrow _safe_call to catch only AttributeError/TypeError/ValueError

HIGH fixes:
- Split 1916-line runtime.py into 4 modules: out_parsers, api_extractors,
  report_render, and slim runtime entry point
- Replace hardcoded "2026" year with regex \d{4} in WGCPJ parser
- Fix bare except Exception patterns with logging and specific types
- Add finally block for COM cleanup in _convert_docx_to_pdf
- Use WD_ALIGN_PARAGRAPH.CENTER instead of magic number 1

MEDIUM fixes:
- Fix inconsistent dict defaults in APIPyInterface generic except path
- Add path traversal check for output_dir parameter
- _convert_bmp_to_png returns None on failure instead of unusable BMP
- Add errors="replace" to subprocess encoding for GBK compatibility
